refactor(wp): move more non wp db functions into Post
