+++ db/migrateWpPostsToArchieMl.ts
@@ -19,8 +19,7 @@ import {
     adjustHeadingLevels,
     findMinimumHeadingLevel,
 } from "./model/Gdoc/htmlToEnriched.js"
-import { getRelatedCharts } from "./wpdb.js"
-import { isPostSlugCitable } from "./model/Post.js"
+import { getPostRelatedCharts, isPostSlugCitable } from "./model/Post.js"
 import { enrichedBlocksToMarkdown } from "./model/Gdoc/enrichedToMarkdown.js"
 
 // slugs from all the linear entries we want to migrate from @edomt
@@ -109,7 +108,7 @@ const migrate = async (): Promise<void> => {
             const text = post.content
             let relatedCharts: RelatedChart[] = []
             if (isEntry) {
-                relatedCharts = await getRelatedCharts(post.id)
+                relatedCharts = await getPostRelatedCharts(post.id)
             }
 
             const shouldIncludeMaxAsAuthor = await isPostSlugCitable(post.slug)
